3.3-permissions/api_with_restrictions/requirements.txt
django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-filter==23.5
